refactor(demo tests): eliminate sleep in tests, reduce fixture redundancy (DEMO-4.11)
- Add DEFAULT_WAIT_SECONDS = 1.0 module-level constant to vultron/demo/utils.py
  so tests can set it to 0.0 without touching production defaults
- Update post_to_inbox_and_wait to use DEFAULT_WAIT_SECONDS when wait_seconds
  is not explicitly provided
- Remove explicit wait_seconds=2.0 overrides in receive_report_demo.py
  (TestClient processes background tasks synchronously, so no sleep needed)
- Create test/demo/conftest.py: sets DEFAULT_WAIT_SECONDS=0.0 for all demo
  tests, provides shared module-scoped client fixture, re-exports
  make_testclient_call helper
- Create test/demo/_helpers.py: make_testclient_call() factory eliminates the
  30-line testclient_call closure duplicated across all 12 demo test files
- Refactor all 12 test/demo/test_*_demo.py fixtures to remove duplicate
  client fixture definitions and use make_testclient_call

Result: demo test suite drops from ~4+ minutes to ~28s total (10x speedup).
Slowest individual test goes from 15.5s to 0.4s.
